moved polyline analysis to utils module + cleanup
