+++ src/compas_timber/utils/__init__.py
@@ -3,6 +3,7 @@ from math import fabs
 from compas.geometry import Plane
 from compas.geometry import Point
 from compas.geometry import add_vectors
+from compas.geometry import angle_vectors_signed
 from compas.geometry import cross_vectors
 from compas.geometry import distance_point_point
 from compas.geometry import dot_vectors
@@ -12,6 +13,7 @@ from compas.geometry import scale_vector
 from compas.geometry import subtract_vectors
 from compas.geometry import Frame
 from compas.geometry import Transformation
+from compas.geometry import Vector
 from compas.geometry import intersection_line_plane
 
 
@@ -171,4 +173,89 @@ def intersection_line_beam_param(line, beam, ignore_ends=False):
     return [Point(*coords) for coords in pts], ref_side_indices
 
 
-__all__ = ["intersection_line_line_param", "intersection_line_plane_param", "intersection_line_beam_param"]
+def _split_into_consecutive_sequences(source, wrap_on):
+    # type: (list[int], int) -> list[list[int]]
+    if not source:
+        return []
+
+    sequences = []
+    current_sequence = [source[0]]
+
+    for i in range(1, len(source)):
+        curr_val = source[i]
+        prev_incremented = (source[i - 1] + 1) % wrap_on
+        if curr_val == prev_incremented:
+            current_sequence.append(curr_val)
+        else:
+            sequences.append(current_sequence)
+            current_sequence = [curr_val]
+
+    sequences.append(current_sequence)  # add the last sequence
+    return sequences
+
+
+def classify_polyline_segments(polyline, normal, direction="cw"):
+    """Classify polyline segments as external or part of a cutout based on turning angles.
+
+    Parameters
+    ----------
+    polyline : :class:`compas.geometry.Polyline`
+        The polyline to classify.
+    normal : :class:`compas.geometry.Vector`
+        The normal vector of the wall. Used as reference palne for turning angles calculation.
+    direction : str, optional
+        The winding direction of the outline around the given normal vector. One of: ("cw", "ccw"). Default is "cw".
+
+    Returns
+    -------
+    tuple(list[int], list[list[int]])
+        A tuple containing two lists:
+        - the first list contains the indices of outline vertices
+        - the second list contains the indices of internal vertices grouped in sequences
+    """
+    if direction not in ("cw", "ccw"):
+        raise ValueError("Direction must be either 'cw' or 'ccw'.")
+
+    # iterate on polyline without p[0] == p[-1]
+    if polyline[0] == polyline[-1]:
+        polyline = polyline[:-1]
+
+    outline_vertices = []
+    internal_vertices = []
+
+    num_points = len(polyline)
+
+    for i in range(num_points):
+        p_prev = polyline[i]
+        p_curr = polyline[(i + 1) % num_points]
+        p_next = polyline[(i + 2) % num_points]
+
+        v1 = Vector.from_start_end(p_prev, p_curr)
+        v2 = Vector.from_start_end(p_curr, p_next)
+
+        angle = angle_vectors_signed(v1, v2, normal, deg=True)
+
+        if direction == "ccw":
+            angle = -angle
+
+        if angle < 0:
+            outline_vertices.append((i + 1) % num_points)
+        else:
+            internal_vertices.append((i + 1) % num_points)
+
+    # vertices that lie on the outline but are at openings count as internal
+    # they are removed from the outline list and added to the internal list
+    internal_groups = _split_into_consecutive_sequences(internal_vertices, wrap_on=num_points)
+    polyline_indices = list(range(num_points))
+    for group in internal_groups:
+        prev_value = polyline_indices[group[0] - 1]
+        next_value = polyline_indices[(group[1] + 1) % num_points]
+        group.insert(0, prev_value)
+        group.append(next_value)
+        outline_vertices.remove(prev_value)
+        outline_vertices.remove(next_value)
+
+    return outline_vertices, internal_groups
+
+
+__all__ = ["intersection_line_line_param", "intersection_line_plane_param", "intersection_line_beam_param", "classify_polyline_segments"]
